dyslexia-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose[cryptography]
psycopg2-binary
alembic
python-dotenv
python-Levenshtein
pytest
httpx
pydantic
groq
transformers
torch
Pillow
python-multipart
opencv-python
numpy
rapidfuzz
python-doctr
